+++ internal/commands/vaultsecrets/vault_secrets.go
@@ -5,7 +5,7 @@ package vaultsecrets
 
 import (
 	"github.com/hashicorp/hcp/internal/commands/vaultsecrets/apps"
-	"github.com/hashicorp/hcp/internal/commands/vaultsecrets/gateway_pools"
+	"github.com/hashicorp/hcp/internal/commands/vaultsecrets/gatewaypools"
 	"github.com/hashicorp/hcp/internal/commands/vaultsecrets/run"
 	"github.com/hashicorp/hcp/internal/commands/vaultsecrets/secrets"
 	"github.com/hashicorp/hcp/internal/pkg/cmd"
@@ -27,7 +27,7 @@ func NewCmdVaultSecrets(ctx *cmd.Context) *cmd.Command {
 
 	cmd.AddChild(apps.NewCmdApps(ctx))
 	cmd.AddChild(secrets.NewCmdSecrets(ctx))
-	cmd.AddChild(gateway_pools.NewCmdGatewayPools(ctx))
+	cmd.AddChild(gatewaypools.NewCmdGatewayPools(ctx))
 	cmd.AddChild(run.NewCmdRun(ctx, nil))
 	return cmd
 }
